requirements.txt
tensorflow
streamlit
pillow
numpy
matplotlib
scipy